test(validation): split the gate's failure-mode tests under the size ceiling
The review fixes pushed test_check_serena_memory_worktree_scope.py to 643
lines, past the 500-line taste ceiling, which moved the taste count ratchet
from 566 to 567. A ratchet baseline may only fall, so the count has to come
down rather than the ceiling go up.

Splitting is the repository's own direction for this: issue #5517 exists to
split an oversized test file rather than waive it, and a `taste-lint: ignore`
here would spend a waiver on growth this branch caused.

The new module takes the cases where the scan cannot run (git raising before
it starts, a bare repository, the argv contract, and the exit code of a run
that examined nothing). The original keeps parsing, reporting, and the happy
path. 482 and 213 lines. The boundary is a size constraint, not a claim that
these are separate concerns, and both headers say so.

Refs #5061
